Get Measurement Chain Data on KWB Cloud

Get Measurement Chain Data on KWB Cloud

Usage

get_measurement_chain_data_on_cloud(dbg = TRUE)

Arguments

dbg

logical indicating whether or not to show debug messages

Value

data frame with the content of "mc_data.zip" in the GeoSalz project folder on the Nextcloud server. The SFTP paths to the files from which the data in "mc_data.zip" originate are returned in attribute "sftp_paths". If either of the files "mc_data.zip" or "mc_files.csv" does not exist, NULL is returned.